Updated feature 3, included version for dryrun

+++ backend.py
@@ -9,6 +9,7 @@ import io
 import os
 #pip install pymupdf
 import fitz
+import ast
 from fpdf import FPDF
 import lorem
 import matplotlib.pyplot as plt
@@ -23,9 +24,10 @@ def call (data):
     # prepare data 
     pre_process_data(data)
     #generate_summary()
-    find_commonalities_and_differences()
+    df_table = find_commonalities_and_differences()
+    highlight_pdf(df_table, True)
     create_pdf()
-    
+
 def pre_process_data(data):
     # Allowing users to upload multiple files and storing the text in a dictionary
     data_dict = {'filename': [], 'text': []}
@@ -41,7 +43,8 @@ def pre_process_data(data):
         data_dict['text'].append(text)
 
         # clear pdf folder
-        os.remove("pdfs/" + uploaded_file.name)
+        if os.path.exists("pdfs/" + uploaded_file.name):
+            os.remove("pdfs/" + uploaded_file.name)
 
         #save pdf to pdfs folder
         with open("pdfs/" + uploaded_file.name, "wb") as f:
@@ -50,7 +53,7 @@ def pre_process_data(data):
     st.session_state.text_df = pd.DataFrame(data_dict)
     st.session_state.text1 = st.session_state.text_df['text'].tolist()[0]
     st.session_state.text2 = st.session_state.text_df['text'].tolist()[1]
-    
+
 
 
 ### Feature 1: Generate Summary of the events
@@ -62,7 +65,7 @@ def generate_summary():
         model = "gpt-3.5-turbo",
         messages=[
         {"role": "system", "content": "Du bist eine Richterassistentin namens Jasmin. Deine Aufgabe ist es, den Richter zu unterstützen, um ihn effizienter bei der Vorbereitung einer Gerichtsverhandlung zu machen."},
-        {"role": "system", "content": "Du erhältst zwei Schriftsätze vom Kläger und vom Beklagten. Du möchtest herausfinden, was die Gemeinsamkeiten und Unterschiede zwischen den beiden Schriftsätzen sind. Du möchtest die wichtigsten Hauptargumente beider Parteien zusammenfassen und eine zusammenhängende Darstellung der rechtlichen Auseinandersetzung liefern."}, 
+        {"role": "system", "content": "Du erhältst zwei Schriftsätze vom Kläger und vom Beklagten. Du möchtest herausfinden, was die Gemeinsamkeiten und Unterschiede zwischen den beiden Schriftsätzen sind. Du möchtest die wichtigsten Hauptargumente beider Parteien zusammenfassen und eine zusammenhängende Darstellung der rechtlichen Auseinandersetzung liefern."},
         {"role": "system", "content": "Bitte fassen Sie den Inhalt der Schriftsätze in einer prägnanten, verständlichen Form zusammen."},
         {"role": "user", "content": "Text 1" + st.session_state.text1},
         {"role": "user", "content": "Text 2" + st.session_state.text2},
@@ -70,8 +73,9 @@ def generate_summary():
 
     # Extracting the generated summary from the api answer
     generated_summary = response['choices'][0]['message']['content']
-    
+
     st.session_state.event_summary = generated_summary
+
 
 ### Feature 2: Overview of disputed and undisputed facts
 
@@ -114,68 +118,229 @@ def find_commonalities_and_differences():
     print("df drop", df.drop(df.index[0], inplace=True))
     st.session_state.fact_table = df
 
+    return df
+
+
+# read out pdfs (double work, should be removed in the end)
+def read_pdf(pdf_path):
+    pdf = fitz.open(pdf_path)
+
+    full_text = ""
+
+    for page_number in range(pdf.page_count):
+        page = pdf[page_number]
+        text = page.get_text()
+        full_text += text
+
+    pdf.close()
+
+    return full_text
 
 
 ### Feature 3: Compare documents by highlighting the briefs
 # input: text to be highlighted (as list), plaintiff/ defendant as context
-def highlight_pdf(highlighted_text, context):
+def highlight_pdf(df, dryrun):
 
-    if context == "plaintiff":
-        pdf = "brief_plaintiff.pdf"
+    # set list of colors for highlighting
+    highlight_colors = [
+        (0.8, 0.95, 0.8),  # Light Mint Green
+        (0.95, 0.8, 0.8),  # Light Peach
+        (0.8, 0.8, 0.95),  # Light Sky Blue
+        (0.9, 0.95, 0.8),  # Pale Lime
+        (0.8, 0.9, 0.95),  # Light Azure
+        (0.95, 0.8, 0.95),  # Light Lavender
+        (0.9, 0.8, 0.95),  # Light Mauve
+        (0.95, 0.95, 0.8),  # Light Lemon
+        (0.8, 0.95, 0.95),  # Light Cyan
+        (0.95, 0.8, 0.8),  # Light Coral
+        (0.8, 0.8, 0.8),  # Light Gray
+        (0.95, 0.95, 0.95),  # Light Silver
+        (0.9, 0.9, 0.8),  # Pale Yellow
+        (0.8, 0.9, 0.9),  # Light Turquoise
+        (0.9, 0.8, 0.9),  # Light Orchid
+        (0.8, 0.9, 0.8),  # Light Sage
+        (0.9, 0.8, 0.8),  # Light Salmon
+        (0.8, 0.8, 0.9),  # Light Periwinkle
+        (0.9, 0.9, 0.9),  # Light Ivory
+        (0.95, 0.7, 0.7),  # Light Salmon (variation)
+    ]
+
+    if dryrun is True:
+
+        # highlight plaintiff
+        pdfIn_plaintiff = fitz.open("pdfs/brief_plaintiff.pdf")
+
+        for page in pdfIn_plaintiff:
+
+            list_plaintiff = ["Der Kläger hat das Motorrad am 8. März 1972 gekauft und erworben. Er ist Eigentü",
+                              "mer des Motorrads.",
+                              "Der Beklagte ist Besitzer des Fahrzeugs. Mit Schreiben vom 3. Juli 2021 wurde der Beklagte zur Herausgabe des Fahrzeugs aufgefordert.",
+                              "n.a.",
+                              "Wie der Beklagte vorprozessual mitgeteilt hat, wurde dieser Fuchsschwanz am 21. Mai 2021 beim Vorbeifahren durch den Radfahrer Matthias Hoster schuldhaft beschädigt.",
+                              "Der Beklagte schuldet nach § 985 BGB Herausgabe des Motorrads. Zum Besitz ist er nicht berechtigt.",
+                              "Da der Beklagte den Betrag von 70,- € aufgrund der Beschädigung des Fuchs",
+                              "schwanzes erhalten hat, ist er zur Zahlung dieses Betrages an den Kläger verpflich",
+                              "tet."]
+
+            # find coordinates of text that should be highlighted
+            text_instances = [page.search_for(text) for text in list_plaintiff]
+
+            j = 0
+            k = [0, 0, 1, 2, 3, 4, 5, 5, 5]
+            # iterate through each instance for highlighting
+            for inst in text_instances:
+                annot = page.add_highlight_annot(inst)
+                annot.set_colors(stroke=highlight_colors[k[j]])
+                # annot.set_info(title=context[k])
+                annot.update()
+                j += 1
+
+        # Saving the PDF Output
+        pdfIn_plaintiff.save("pdfs/brief_plaintiff_highlighted.pdf")
+        print("PDF Plaintiff stored")
+
+        # highlight defendant
+        pdfIn_defendant = fitz.open("pdfs/brief_defendant.pdf")
+
+        for page in pdfIn_defendant:
+
+            list_defendant = [
+                "Der Beklagte ist Besitzer des Motorrads. Es wird jedoch ausdrücklich bestritten, dass der Kläger Eigentü",
+                "mer ist.", "Von diesem erwarb der Beklagte das Motorrad mit dem angebundenen Fuchs",
+                "schwanz am 10. April 2021 redlich für 600,- € zum Zwecke der Restaurierung. Auch die Zulassungsbescheinigung Teil II wurde ihm übergeben.",
+                "Das Motorrad war vor der Restaurierung nicht verkehrstüchtig, trotz seines Alters waren offensichtlich seit längerem keine Teile erneuert worden. Auch der äußere Zu",
+                "stand war reparaturbedürftig.", "n.a.",
+                "Die Klage ist abzuweisen, weil der Kläger nicht Eigentümer des Motorrads ist.", "n.a."]
+
+            # find coordinates of text that should be highlighted
+            text_instances = [page.search_for(text) for text in list_defendant]
+
+            j = 0
+            k = [0, 0, 1, 1, 2, 2, 3, 4, 5]
+            # iterate through each instance for highlighting
+            for inst in text_instances:
+                annot = page.add_highlight_annot(inst)
+                annot.set_colors(stroke=highlight_colors[k[j]])
+                # annot.set_info(title=context[k])
+                annot.update()
+                j += 1
+
+        # Saving the PDF Output
+        pdfIn_defendant.save("pdfs/brief_defendant_highlighted.pdf")
+        print("PDF Defendant stored")
+
     else:
-        pdf = "brief_defendant.pdf"
+        # get text form (need to be sure to have the correct document)
+        full_text_plaintiff = read_pdf("pdfs/brief_plaintiff.pdf")
+        full_text_defendant = read_pdf("pdfs/brief_defendant.pdf")
 
-    pdfIn = fitz.open(pdf)
+        # list_contexts = df.iloc[1:, 0].tolist()
+        list_facts_plaintiff = df.iloc[1:, 2].tolist()
+        list_facts_defendant = df.iloc[1:, 3].tolist()
 
-    # highlighted_text = ["Das Motorrad war vor der Restaurierung nicht verkehrstüchtig, trotz seines Alters waren "
-    #                    "offensichtlich seit längerem keine Teile erneuert worden.",
-    #                    "Der Beklagte erfuhr erstmals bei einem Gespräch mit Herrn Grünbaum am 5. Juni 2021, dass "
-    #                    "dieser das Motorrad von einer unbekannten Person erworben hatte und dass dieses zuvor dem
-    #                    Sohn des Klägers gestohlen worden war."]
+        string_facts_plaintiff = ''.join(map(str, list_facts_plaintiff))
+        string_facts_defendant = ''.join(map(str, list_facts_defendant))
 
-    for page in pdfIn:
+        # find the original plaintiff text
+        response = openai.ChatCompletion.create(
+            model="gpt-3.5-turbo",
+            messages=[
+                {"role": "system",
+                 "content": "Du bist eine Richterassistentin namens Jasmin. Deine Aufgabe ist es, den Richter zu unterstützen, um ihn effizienter bei der Vorbereitung einer Gerichtsverhandlung zu machen."},
+                {"role": "system",
+                 "content": "Du erhältst einen originalen Schriftsatz von einem Kläger sowie eine Liste mit Stichpunkten, welche einen Fakt innerhalb des Dokuments beschreiben. Du möchtest die Textstellen im originalen Dokument suchen, welche von dem jeweiligenn Stichpunkt beschrieben wird."},
+                {"role": "system",
+                 "content": "Gibt bitte dir originale Textstellen in einer einzigen Liste zurück und zwar in der gleichen Reihenfolge wie die Stichpunkte. Bitte nimm keine ganzen Absätze, sondern nur kurze originale Textstücke. Achte Beispiel: ['Originaltext1', 'Originaltext2', ...]."},
+                {"role": "user", "content": full_text_plaintiff},
+                {"role": "user", "content": string_facts_plaintiff},
+            ])
+        text_plaintiff = response['choices'][0]['message']['content']
 
-        # find coordinates of text that should be highlighted
-        text_instances = [page.search_for(text) for text in highlighted_text]
+        print("ChatGPT plaintiff")
+        print(text_plaintiff)
 
-        # set list of colors for highlighting
-        highlight_colors = [
-            (1, 0, 0),  # Red
-            (0, 1, 0),  # Green
-            (0, 0, 1),  # Blue
-            (1, 1, 0),  # Yellow
-            (1, 0, 1),  # Magenta
-            (0, 1, 1),  # Cyan
-            (0.5, 0, 0),  # Maroon
-            (0, 0.5, 0),  # Olive
-            (0, 0, 0.5),  # Navy
-            (0.5, 0.5, 0),  # Olive Green
-            (0.5, 0, 0.5),  # Purple
-            (0, 0.5, 0.5),  # Teal
-            (0.8, 0.4, 0),  # Dark Orange
-            (0.8, 0, 0.4),  # Dark Reddish Pink
-            (0.4, 0.8, 0),  # Dark Lime Green
-            (0, 0.8, 0.4),  # Dark Mint Green
-            (0.4, 0, 0.8),  # Dark Indigo
-            (0, 0.4, 0.8),  # Dark Sky Blue
-            (0.7, 0.7, 0.7),  # Light Gray
-            (0.4, 0.4, 0.4),  # Medium Gray
-            (0.9, 0.9, 0),  # Pale Yellow
-            (0, 0.9, 0.9),  # Light Cyan
-            (0.9, 0, 0.9),  # Pink
-            (0.7, 0.7, 0.9),  # Light Purple
-            (0.9, 0.7, 0.7),  # Light Salmon
-        ]
-        # iterate through each instance for highlighting
-        i = 0
-        for inst in text_instances:
-            annot = page.add_highlight_annot(inst)
-            annot.set_colors(stroke=highlight_colors[i])
-            annot.update()
-            i += 1
+        # Find the position of '[' and ']'
+        start_index = text_plaintiff.find('[')
+        end_index = text_plaintiff.find(']')
 
-    # Saving the PDF Output
-    pdfIn.save("brief_" + context + "_highlighted.pdf")
+        # Extract the substring containing the list representation
+        list_representation = text_plaintiff[start_index:end_index + 1]
+
+        try:
+            list_plaintiff = ast.literal_eval(list_representation)
+        except (ValueError, SyntaxError) as e:
+            print(f"Error: {e}")
+
+
+
+        # find the original defendant text
+        response = openai.ChatCompletion.create(
+            model="gpt-3.5-turbo",
+            messages=[
+                {"role": "system",
+                 "content": "Du bist eine Richterassistentin namens Jasmin. Deine Aufgabe ist es, den Richter zu unterstützen, um ihn effizienter bei der Vorbereitung einer Gerichtsverhandlung zu machen."},
+                {"role": "system",
+                 "content": "Du erhältst einen originalen Schriftsatz von einem Kläger sowie eine Liste mit Stichpunkten, welche einen Fakt innerhalb des Dokuments beschreiben. Du möchtest die Textstellen im originalen Dokument suchen, welche von dem jeweiligenn Stichpunkt beschrieben wird."},
+                {"role": "system",
+                 "content": "Gibt bitte dir originale Textstellen in einer einzigen Liste zurück und zwar in der gleichen Reihenfolge wie die Stichpunkte. Bitte nimm keine ganzen Absätze, sondern nur kurze originale Textstücke. Beispiel: ['Originaltext1', 'Originaltext2', ...]."},
+                {"role": "user", "content": full_text_defendant},
+                {"role": "user", "content": string_facts_defendant},
+            ])
+        text_defendant = response['choices'][0]['message']['content']
+
+        # Find the position of '[' and ']'
+        start_index = text_defendant.find('[')
+        end_index = text_defendant.find(']')
+
+        # Extract the substring containing the list representation
+        list_representation = text_defendant[start_index:end_index + 1]
+
+        try:
+            list_defendant = ast.literal_eval(list_representation)
+        except (ValueError, SyntaxError) as e:
+            print(f"Error: {e}")
+
+        # highlight plaintiff
+        pdfIn_plaintiff = fitz.open("pdfs/brief_plaintiff.pdf")
+
+        for page in pdfIn_plaintiff:
+
+            # find coordinates of text that should be highlighted
+            text_instances = [page.search_for(text) for text in list_plaintiff]
+
+            i = 0
+            # iterate through each instance for highlighting
+            for inst in text_instances:
+                annot = page.add_highlight_annot(inst)
+                annot.set_colors(stroke=highlight_colors[i])
+                # annot.set_info(title=list_contexts[i])
+                annot.update()
+                i += 1
+
+        # Saving the PDF Output
+        pdfIn_plaintiff.save("pdfs/brief_plaintiff_highlighted.pdf")
+        print("PDF Plaintiff stored")
+
+        # highlight defendant
+        pdfIn_defendant = fitz.open("pdfs/brief_defendant.pdf")
+
+        for page in pdfIn_defendant:
+
+            # find coordinates of text that should be highlighted
+            text_instances = [page.search_for(text) for text in list_defendant]
+
+            i = 0
+            # iterate through each instance for highlighting
+            for inst in text_instances:
+                annot = page.add_highlight_annot(inst)
+                annot.set_colors(stroke=highlight_colors[i])
+                # annot.set_info(title=list_contexts[i])
+                annot.update()
+                i += 1
+
+        # Saving the PDF Output
+        pdfIn_defendant.save("pdfs/brief_defendant_highlighted.pdf")
+        print("PDF Defendant stored")
 
 
 ### Chatbot Feature
@@ -210,7 +375,7 @@ def generate_response(prompt):
 
 def create_pdf():
     text = st.session_state.event_summary
-    df = st.session_state.fact_table 
+    df = st.session_state.fact_table
     #df = pd.read_excel("fact_table.xlsx")
     #use first column as index and drop it
     #df.set_index(df.columns[0], inplace=True)
@@ -256,10 +421,10 @@ def create_pdf():
     pdf.add_page(orientation='L')
 
     pdf.image('mytable.png', x = 0, y = 0, w = 300, h = 200)
-  
+
     #pdf.output("output.pdf")
 
-    
+
     # Save the PDF to a bytes object
     #pdf_out = io.BytesIO()
     #pdf.output(dest='S')
